feat: Add Dockerfiles and configuration files for main NX and sub web apps
This commit includes the addition of Dockerfiles for the main NX workspace and the sub web apps, frontend and backend. It also adds the necessary Docker configuration files, such as dockerignore and docker-compose for running containers on Docker. Additionally, an nginx.conf file has been added for launching on nginx.